관리자페이지 대시보드 그래프 크기 수정 + 이벤트페이지 수정

+++ shoekream/src/main/java/com/supreme/shoekream/service/MemberApiLogicService.java
@@ -93,7 +93,7 @@ public class MemberApiLogicService extends BaseService<MemberApiRequest, MemberA
 //}
     public Boolean point( Long memberidx){
         Member member = memberRepository.findById(memberidx).get();
-        if(member.getPoint()>=1000){
+        if(member.getPoint().intValue() >= 1000){
             return true;
         }else{
             return false;
